requirements.txt
numpy
opencv-python
pandas
streamlit
pathlib
pillow
dataclasses
ultralytics
playsound